// src/app/models/allergy.ts
export class Allergy {
	Allergy_Id: number = 0;
	Allergy_Code: string = "";
	Allergy_Type: string = "";
	Allergy_Name: string = "";
	Allergy_Description: string = "";
	Allergy_Clinical_Information: string = "";
	Allergy_Source: string = "";
	allerginicity: string = "";
	allergy_IsFatal: string = "";

	// constructor(Allergy_Code = "", Allergy_Type = "", Allergy_Name = "", Allergy_Description = "", Allergy_Clinical_Information = "", allergy_IsFatal = "") {
	// 	this.Allergy_Code = Allergy_Code;
	// 	this.Allergy_Type = Allergy_Type;
	// 	this.Allergy_Name = Allergy_Name;
	// 	this.Allergy_Description = Allergy_Description;
	// 	this.Allergy_Clinical_Information = Allergy_Clinical_Information;
	// 	this.allergy_IsFatal = allergy_IsFatal;
	// }
}